Do required changes
