+++ .eslintrc.json
@@ -18,9 +18,7 @@
     "react/no-unescaped-entities": 0,
     "react/jsx-filename-extension": ["warn", { "extensions": [".js", ".jsx"] }],
     "react/react-in-jsx-scope": "off",
-    "import/no-unresolved": "off",
-    "no-shadow": "off",
-    "no-multiple-empty-lines": "off"
+    "import/no-unresolved": "off"
   },
   "ignorePatterns": ["dist/", "build/"]
 }
