+++ src/API/spacedata.js
@@ -1,0 +1,5 @@
+const spaceData = {
+  missions: 'https://api.spacexdata.com/v3/missions',
+};
+
+export default spaceData;
