--- src/redux/missions/middlewares.js
@@ -1,12 +1,0 @@
-import { createStore, combineReducers, applyMiddleware } from 'redux';
-import thunk from 'redux-thunk';
-import logger from 'redux-logger';
-import spaceReducer from './missions';
-
-const reducer = combineReducers({
-  spaceReducer,
-});
-
-const store = createStore(reducer, applyMiddleware(thunk, logger));
-
-export default store;
